requirements.txt
django==3.1.1
djangorestframework-jwt==1.11.0
djangorestframework==3.11.1
drf_yasg==1.17.1
psycopg2-binary==2.8.6
django-silk==4.1.0
python-decouple==3.3